+++ changelog.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file. This project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## v4.1.2 - _2022-11-29_
+
+v4.1.2 Fix for query large dataset client usage
+
+### Fixes
+- Fix query large dataset client usage
+
 ## v4.1.1 - _2022-11-23_
 
 v4.1.1 fixes a few issues with the schema creation UI